Messing about with re-arranging the dashboard code, probably won't work
I moved the dashboard code to it's own class. I needed to make all the methods and variables in the class static to allow it to be called from the robot class. By calling it from the Robot class it opens the dashboard up to putting whatever it wants on whatever tab because it's all at the highest level. I don't know if the robotPeriodic() method is called often enough to make this work so I'll just have to try it when no-one else is using the robot.